notion-clone-python-sveltekit/backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6
psycopg2==2.9.9
python-dotenv==1.0.1
sqlalchemy==2.0.28
alembic==1.13.1
uuid==1.30
eventlet==0.35.2
y-py==0.6.2
ypy-websocket==0.12.4 